Add account page; Add 404; Rename handlers
Added some content to the account page.  There is a form, but it doesn't
do anything yet.  The other data for the page is also not populated yet
(votes, etc).

All the handler_* functions have been renamed to remove the underscore.

A proper 404 page has been added and can be used with a call to
Server.do404().  This will display a page with an error and send the
correct HTTP status code. Might extend this later for use with other
satus codes.

Added the page title to the header div.  It is put next to the
"MoviePolls" link, but it does not link itself.