Implement disabling voting
Implemented the VotingEnabled configuration option.  With this set,
voting works as expected.  However with it unset the links to vote on
selections are removed and an error is displayed if the user tries to
vote directly with the vote URL.

The UI to enable and disable this option has also been added to the
admin config page.

Some smaller changes have been made that remove explicit structures for
template data in favor of anonymous structs in the calling method.