Rework the cycle and movie relationship a bit; implement history page
The cycles and movies relationship has been modififed slightly.  Each
cycle now has a list of movies that were watched during it (only filled
when returned from GetPastCycles().  This may change later).
Additionally, each movie has a pointer to the cycle it was watched in
instead of an end date.

The /history page has been implemented.  It currently shows the past
five cycles and the movies watched during them.