Implement deleting and purging a user
Deleting a user does not remove the account or the votes.  Instead, it
changes the account's name to [deleted] and removes all of its info
(password, email, oauth, and notification settings).  The password date
is also changed which will logout the user if they have an active
session.

Purging a user completely removes the user's account as well as the
votes.  Currently, this will remove *all* the users votes, including
those on movies that have already been watched.  This behavior may be
tweaked in the future.

This is step one for #31.