Fix vote decay in JSON connector (incl. test)
Vote decaying is not connected to the rest of the application yet.  This
is just fixing the mechanism for it.  Currently, only the JSON connector
implements DecayVotes().

The test case simulates 4 cycles with 2 movies being added to each.  One
movie is marked as watched each cycle while the other is a decay
candidate.  The test case decays votes older than two cycles and
verifies that only the correct votes were removed.