Move CSS files into /static/css/; Fix static file serving
Fixed the way paths are determined when serving static files (in both
"static/" and "posters/").  Instead of just taking the base name of a
file, run the full path through filepath.Clean() and use that.  It
*shouldn't* be possible to get arbitrary files on the server.

Moved the CSS files into /static/css because that makes more sense than
having them in just /static.