Implement generic URL/Key pairs for password resets
The URL/Key pair that was used for claiming admin has been expanded to
also be used for password resets.  The two password fields have been
removed from the edit user page in the admin panel and have been
replaced with a link to generate a password reset link.

Eventually, this link will be sent to an email (if set).

This is the second part of #31.