Implement removing a movie
This is an admin and mod only feature.  Only active movies can be
removed.  If a movie has any votes, they will be removed first.

It may be a good idea to allow the user that added it the ability to
remove the movie, so long as it has no votes.