Adapted the movie-info page to the new link struct
This way to display links assumes that the first link is the source link
(what should be assured by the logic of the server).
One could change the "logic" in the template to fix that assumption but
i don't fancy too messy templates since a rework of the visuals is still
pending.