Added new function ValidateLink and DetermineLinkType rework
ValidateLink prepends the https protocol "hint" when no protocol
seperator "//" is found in the struct on which the function is called.
This function was necessary since i.e. "foobar.com" in the link input
resulted on a functioning link that linked to localhost:8090/foobar.com
when clicked on the frontend.

I also reworked the DetermineLinkType function to not return the type as
string and instead set it as "sideeffect" -> works more like a setter
now.