Tried to make sense of the AddMovie function stub
I **ASSUME** input fields are supposed to have the value of the field,
hence i created a new function to parse these values into a new struct
to be used later on in the logic.

Parsing of links themselfs is also extracted from the general parsing
function to enable adapting the logic later.