Adapted `AddMovie` to a new structure
All the input validation logic was moved to `validateForm` which takes
and returns a map of `InputField`.
This map initially only holds the values for the fields of the form, but
after the validation the `Error` field of these `InputField` structs may
be filled with errors detected during validation.

Also `AddMovie` no longer returns an error, since all errors are
contained in the map of `InputField` anyways.