Added (for now hardcoded) port to the host string
Now when the new admin authentication is generated for the first time,
the port is appended and saved in the db, hence all following db calls
for host also includes the port (this finally also adds valid links to
the password reset).

Added a new function `SetHostAddress` for this purpose.