Added new struct callbackError to server struct
This new callbackError struct holds the user id and the message for the
error.
This struct will have an empty message if no error occurs and will be
unused if this is the case.
If the message is not empty, it is rendered on the account page IF the
user id matches.