Making use and clearing callbackError struct
If the message of the struct is not empty we check the user id.
If the user id matches we save the message in the data passed to the
template and clear out the callbackError struct.
If either the message is empty or the id does not match we ignore the
struct.